+++ src/Bucket.ts
@@ -8,10 +8,19 @@ import {LabelMap, ReadableRecord, WritableRecord} from "./Record";
 /**
  * Options for querying records
  */
-export interface QueryOptions  {
+export interface QueryOptions {
     ttl?: number;   // Time to live in seconds
     include?: LabelMap; //  include only record which have all these labels with the same value
     exclude?: LabelMap;  //  exclude record which have all these labels with the same value
+}
+
+/**
+ * Options for writing records
+ */
+export interface WriteOptions {
+    ts?: bigint;    // timestamp of the record
+    labels?: LabelMap;  // labels of the record
+    contentType?: string;  // content types of the record
 }
 
 /**
@@ -85,17 +94,25 @@ export class Bucket {
     /**
      * Start writing a record into an entry
      * @param entry name of the entry
-     * @param ts {BigInt} timestamp in microseconds for the record. It is current time if undefined.
-     * @param labels {Record<string, LabelMap} labels for the record, should be a key-value map
+     * @param options {BigInt | WriteOptions} timestamp in microseconds for the record or options. It is current time if undefined.
      * @return Promise<WritableRecord>
      * @example
      * const record = await bucket.beginWrite("entry", 1203121n, {label1: "value1", label2: "value2"});
      * await record.write("Hello!);
      */
-    async beginWrite(entry: string, ts?: bigint, labels?: LabelMap): Promise<WritableRecord> {
-        ts ||= BigInt(Date.now() * 1000);
+    async beginWrite(entry: string, options?: bigint | WriteOptions): Promise<WritableRecord> {
+        let localOptions: WriteOptions = {};
+        if (options !== undefined) {
+            if (typeof (options) === "bigint") {
+                localOptions = {ts: options};
+            } else {
+                localOptions = options as WriteOptions;
+            }
+        }
+        localOptions.ts = localOptions.ts ?? BigInt(Date.now()) * 1000n;
 
-        return Promise.resolve(new WritableRecord(this.name, entry, ts, this.httpClient, labels ? labels : {}));
+
+        return Promise.resolve(new WritableRecord(this.name, entry, localOptions, this.httpClient));
     }
 
     /**
@@ -189,7 +206,8 @@ export class Bucket {
         return new ReadableRecord(BigInt(headers["x-reduct-time"]), BigInt(headers["content-length"]),
             headers["x-reduct-last"] == "1",
             data,
-            labels);
+            labels,
+            headers["content-type"],);
     }
 
 }